+++ test/src/unit-capi-incomplete.cc
@@ -1,5 +1,5 @@
 /**
- * @file   unit-capi-consolidation.cc
+ * @file   unit-capi-incomplete.cc
  *
  * @section LICENSE
  *